Adding a second reactive variable that tracks the URL with query string
The existing one just tracks the basic router string, which doesn't
allow us to make a page respond dynamically to data included in the
link itself.

Ultimately this still feels pretty hacky, though. It should be possible
to do this without the hidden input field, using Javascript variables
or Shiny events or something.